Lab 10: Progress so far

+++ Lab-10/Sketchpad/src/main/java/Circle.java
@@ -4,9 +4,9 @@ public class Circle extends Shape{
     }
 
     @Override
-    public double area(){
+    public int area(){
         double a = 3.1416 * height * width;
-        return a;
+        return (int)a;
     }
 
     @Override
